"Refactor the document replacement modal
The code for the document replacement modal has been refactored to improve reusability and consistency. Constants and action creators specific to project summary documents have been swapped out for more generic ones. Also, the alert message warning users about the implications of replacing files is now dynamic, allowing it to be rendered based on the specific context."
