+++ services/core-web/src/components/Forms/projectSummaries/ProjectSummaryDocumentUpload.js
@@ -87,13 +87,9 @@ export class ProjectSummaryDocumentUpload extends Component {
                 {
                   ...doc,
                   key: doc.mine_document_guid,
-                  mine_document_guid: doc.mine_document_guid,
-                  document_manager_guid: doc.document_manager_guid,
-                  document_name: doc.document_name,
                   category: this.props.projectSummaryDocumentTypesHash[
                     doc.project_summary_document_type_code
                   ],
-                  upload_date: doc.upload_date,
                 },
                 ...docs,
               ],
