+++ services/core-web/src/components/common/DocumentTable.tsx
@@ -55,6 +55,7 @@ interface DocumentTableProps {
   fileOperationPermissionMap: { operation: FileOperations; permission: string | boolean }[];
   userRoles: string[];
   handleRowSelectionChange: (arg1: MineDocument[]) => void;
+  replaceAlertMessage?: string;
 }
 
 // eslint-disable-next-line @typescript-eslint/no-shadow
@@ -75,6 +76,7 @@ export const DocumentTable = ({
   closeModal,
   removeDocument,
   openDocument,
+  replaceAlertMessage = "The replaced file will not reviewed as part of the submission.  The new file should be in the same format as the original file.",
   ...props
 }: DocumentTableProps) => {
   const [rowSelection, setRowSelection] = useState([]);
@@ -167,6 +169,7 @@ export const DocumentTable = ({
           setDocuments(newDocuments);
         },
         document: doc,
+        alertMessage: replaceAlertMessage,
       },
       content: modalConfig.REPLACE_DOCUMENT,
     });
