+++ services/minespace-web/common/actionCreators/projectActionCreator.ts
@@ -590,42 +590,3 @@ export const removeDocumentFromProjectDecisionPackage = (
     })
     .finally(() => dispatch(hideLoading()));
 };
-
-export const postNewDocumentVersion = ({
-  mineGuid,
-  mineDocumentGuid,
-  documentManagerVersionGuid,
-}: {
-  mineGuid: string;
-  mineDocumentGuid: string;
-  documentManagerVersionGuid: string;
-}): AppThunk<Promise<AxiosResponse<IProjectSummary>>> => (
-  dispatch
-): Promise<AxiosResponse<IProjectSummary>> => {
-  dispatch(request(reducerTypes.POST_NEW_DOCUMENT_VERSION));
-  dispatch(showLoading());
-
-  const payload = { document_manager_version_guid: documentManagerVersionGuid };
-
-  return CustomAxios()
-    .post(
-      `${ENVIRONMENT.apiUrl}/mines/${mineGuid}/documents/${mineDocumentGuid}/versions`,
-      payload,
-      createRequestHeader()
-    )
-    .then((response: AxiosResponse<IProjectSummary>) => {
-      notification.success({
-        message: "Successfully created new document version",
-        duration: 10,
-      });
-      dispatch(success(reducerTypes.POST_NEW_DOCUMENT_VERSION));
-      return response;
-    })
-    .catch((err) => {
-      dispatch(error(reducerTypes.POST_NEW_DOCUMENT_VERSION));
-      throw new Error(err);
-    })
-    .finally(() => {
-      dispatch(hideLoading());
-    });
-};
