+++ services/minespace-web/common/constants/API.js
@@ -137,7 +137,7 @@ export const PROJECT_SUMMARY_DOCUMENTS = ({ projectGuid, projectSummaryGuid, min
     { mine_guid: mineGuid }
   )}`;
 //New file version upload
-export const NEW_VERSION_PROJECT_SUMMARY_DOCUMENTS = ({ mineGuid, mineDocumentGuid }) =>
+export const NEW_VERSION_DOCUMENTS = ({ mineGuid, mineDocumentGuid }) =>
   `/mines/${mineGuid}/documents/${mineDocumentGuid}/versions/upload`;
 export const PROJECT_SUMMARY_DOCUMENT = (
   projectGuid,
